feat: add exceptions on product service

+++ src/main/java/com/ph/ecommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.ph.ecommerce.services;
 import com.ph.ecommerce.dto.ProductDTO;
 import com.ph.ecommerce.entities.Product;
 import com.ph.ecommerce.repositories.ProductRepository;
+import com.ph.ecommerce.services.exceptions.DataBaseException;
 import com.ph.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -45,18 +49,34 @@ public class ProductService {
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
 
-        Product product = repository.getReferenceById(id);
+        try {
+            Product product = repository.getReferenceById(id);
 
-        copyToEntity(product, dto);
+            copyToEntity(product, dto);
 
-        product = repository.save(product);
+            product = repository.save(product);
 
-        return new ProductDTO(product);
+            return new ProductDTO(product);
+        }
+        catch (EntityNotFoundException error) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+           throw  new ResourceNotFoundException("Resource not found");
+        }
+
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException error) {
+            throw new DataBaseException("Integrity violation");
+        }
+
     }
 
     public void copyToEntity(Product product, ProductDTO dto) {
